pyproject.toml
[tool.poetry]
name = "data-validation-framework"
version = "0.1.0"
description = "data validation library to compare and validate two hive tables"
authors = ["mopsde <v-tushara@zillowgroup.com>"]


[[tool.poetry.source]]
name = "artifactory"
url = "https://artifactory.zgtools.net/artifactory/api/pypi/zg-pypi/simple"

[[tool.poetry.source]]
name = "public-pypi"
url = "https://pypi.org/simple"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pyspark = { version = ">=3.3.0", source = "public-pypi" }
pyyaml = "^5.3.1"
contract-registry-client-library =  { version = "2.28+master.x9d42118", source = "artifactory" }
urllib3 = {version = ">=1.0,<2.0", source = "public-pypi"}
requests = {version = "<2.30", source = "public-pypi"}
# ipython = "*"

[tool.poetry.scripts]
generate_table_comparison_details_wheel_task = "data_validation_framework.process_validations_wheel_task:generate_table_comparison_details_wheel_task"
generate_multiple_table_comparison_details_wheel_task = "data_validation_framework.process_validations_wheel_task:generate_multiple_table_comparison_details_wheel_task"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
black = "^23.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
